requirements.txt
fastapi
uvicorn
dnspython
python-jose
passlib
pymongo
